Bug fix
Fixed a bug that was misclassifying some reference or no-call blocks.  If a block was emitted due to being separated from the current position by one or more bases the ref_block variable was reset leaving the current position with no classification.  This led to some cases where the first position in a block did not match the rest of the block.